Created the initial file download of the papers. and written tests for this
Issue.
-----
This is only working in development, because static files are being served by django
It is needed to find a way to make this work even if debug is False,
Options

I think we can use whitenoise for this or use nginx server to handle the serving of media and static files.